requirements.txt
#rest apis
flask
flask_cors
uvicorn
gunicorn
load_dotenv

#postgres db
psycopg2-binary

# image embed
pillow
#torch
#torchvision
#transformers

numpy<2
#CLIP
#scikit-learn
sentence_transformers